Cambio viewModel details y maps
